weather ui

+++ app/src/main/java/sample/network/rahul/android_weather_app/datasource/data/Sys.kt
@@ -20,8 +20,8 @@ class Sys {
     var country: String? = null
     @SerializedName("sunrise")
     @Expose
-    var sunrise: Int? = null
+    var sunrise: Long? = null
     @SerializedName("sunset")
     @Expose
-    var sunset: Int? = null
+    var sunset: Long? = null
 }
